+++ core/structures.js
@@ -134,6 +134,10 @@ export class Structure {
 		};
 		return new Structure(schema, (input, context) => {
 			if (input === undefined) return fallback;
+			if (typeof input === "string") {
+				const parsed = Number.parseFloat(input);
+				if (!Number.isNaN(parsed)) return parsed;
+			}
 			if (typeof input !== "number") {
 				throw new StructError("Not a number", context?.path);
 			}
@@ -159,6 +163,24 @@ export class Structure {
 				throw new StructError("Not a string or URL", context?.path);
 			}
 			return new URL(input);
+		});
+	}
+
+	/**
+	 * @param {boolean} fallback
+	 * @returns {Structure<boolean>}
+	 */
+	static boolean(fallback) {
+		const schema = {
+			type: "boolean",
+			default: fallback,
+		};
+		return new Structure(schema, (input, context) => {
+			if (input === undefined) return fallback;
+			if (typeof input !== "boolean") {
+				throw new StructError("Not a boolean", context?.path);
+			}
+			return input;
 		});
 	}