+++ core/structures.test.js
@@ -1,6 +1,11 @@
-import { assertInstanceOf } from "../bundle/deno/core/test-deps.js";
 import { StructError, Structure } from "./structures.js";
-import { assertEquals, assertThrows, describe, it } from "./test-deps.js";
+import {
+	assertEquals,
+	assertThrows,
+	describe,
+	it,
+	assertInstanceOf,
+} from "./test-deps.js";
 
 describe("StructError", () => {
 	describe("constructor", () => {
@@ -209,6 +214,15 @@ describe("Structure", () => {
 				"should fall back to the default if undefined is passed",
 			);
 		});
+		// TODO: I'm not sure if this should be on Structure or Configuration
+		it("parses string integers", () => {
+			const struct = Structure.number(42);
+			assertEquals(
+				struct.process("33"),
+				33,
+				"should parse the integer out of the string",
+			);
+		});
 		it("throws for non-numbers", () => {
 			const struct = Structure.number(42);
 
@@ -225,6 +239,46 @@ describe("Structure", () => {
 		it("generates JSON schema", () => {
 			const struct = Structure.number(42);
 			assertEquals(struct.schema, { type: "number", default: 42 });
+		});
+	});
+
+	describe("boolean", () => {
+		it("creates a structure", () => {
+			const struct = Structure.boolean(false);
+			assertInstanceOf(struct, Structure);
+		});
+		it("allows booleans", () => {
+			const struct = Structure.boolean(false);
+			assertEquals(
+				struct.process(true),
+				true,
+				"should allow boolean values through",
+			);
+		});
+		it("uses the fallback", () => {
+			const struct = Structure.boolean(false);
+			assertEquals(
+				struct.process(undefined),
+				false,
+				"should fall back to the default if undefined is passed",
+			);
+		});
+		it("throws for non-booleans", () => {
+			const struct = Structure.boolean(false);
+
+			const error = assertThrows(
+				() => struct.process("a string", { path: ["some", "path"] }),
+				StructError,
+			);
+			assertEquals(
+				error,
+				new StructError("Expected a boolean", ["some", "path"]),
+				"should throw a StructError and capture the context",
+			);
+		});
+		it("generates JSON schema", () => {
+			const struct = Structure.boolean(false);
+			assertEquals(struct.schema, { type: "boolean", default: false });
 		});
 	});